requirements.txt
flask
flask-cors
psycopg2-binary
python-dotenv
